MSVC convert C written files (SimpleLog.c, Util.c) into C89
MSVC does not support C99. A fun part of C89 is that variables
must be defined at the start of a block. Heed el standard.

This is a lengthier diff. As I discovered I made at least one error
before creating this patchset, I went triplechecking to make sure
the code is functionally equivalent. Of course I may still be wrong.