MSVC dll installation - Needs support from upstream mingwlibs / vclibs
Install dlls into the binary output directory (ie. spring executable
location). At the moment of writing, the dlls are not included in
vclibs. Lobby the evil vclibs dev into submission to have a go at the
binary deliciousness.

The list of dlls in MinGW mingwlibs (Look in the dll subfolder):
https://github.com/spring/mingwlibs/commit/eba0f0748a29d39fda22178c33e3b98540c6e12e
The list of dlls that should be put in vclibs dllMSVC subfolder is
the same as for MinGW, except that:
  - freetype6 is static in vclibs, it is also the only to require zlib
  - zlib1     is not needed
  - glew32    is static in vclibs
These three libraries do not need to have there dll included.

Finally remember to include Awk.exe - into the vclibs bin subfolder.