MSVC Correct use of CMAKE_PREFIX_PATH with CMAKE_FIND_ROOT_PATH
I've actually reported the issue to CMake bugtracker.
(Issue #14619, was manifesting as a breakage in the FIND_FILE command)

From the CMake developer reply it is seen that Spring's use of
CMAKE_FIND_ROOT_PATH is not the intended way - It worked by accident,
as allowed by a change in CMake made somewhere in year 2007.

"""The CMAKE_FIND_ROOT_PATH is not a list of locations to search
   and was only made to appear to be so by '/' in the
   CMAKE_SYSTEM_PREFIX_PATH."""
"""Use CMAKE_PREFIX_PATH instead"""

Spring's CMAKE_PREFIX_PATH was set to ${MINGWLIBS} before.
It is set to '/' now, cooperating with CMAKE_FIND_ROOT_PATH.